Move migration tests from migration script to scripts/tests
